moved graph routines into separate module
