+++ src/sigmaepsilon/math/graph/__init__.py
@@ -1,1 +1,2 @@
 from .graph import *
+from .utils import *
